DSStest/requirements.txt
asgiref==3.2.3
DateTime==4.3
Django==3.0.3
djangorestframework==3.11.0
joblib==0.14.1
numpy==1.18.1
pandas==1.0.1
patsy==0.5.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.0
statsmodels==0.11.1
zope.interface==4.7.2
